// src/enums/business-type.enum.ts
export enum BusinessType {
  ITEM = 'item',
  BOOK = 'book',
  BOOK_MEMBER = 'bookMember',
  FUND_BOOK = 'fundBook',
  FUND = 'fund',
  CATEGORY = 'category',
  SHOP = 'shop',
  SYMBOL = 'symbol',
  USER = 'user',
  ATTACHMENT = 'attachment',
  NOTE = 'note',
  ROOT = 'root',
  DEBT = 'debt',
  GIFT_CARD = 'giftCard',
  ACTIVITY = 'activity',
  ACTIVITY_DEFINITION = 'activityDefinition',
  VEHICLE = 'vehicle',
  FUEL_RECORD = 'fuelRecord',
  ITEM_RELATION = 'itemRelation',
  USER_SHARE = 'userShare',
  RECURRING_CONFIG = 'recurringConfig',
  BOOKKEEPING_RULE = 'bookkeepingRule',
  PERIOD_CYCLE = 'periodCycle',
  PERIOD_DAILY_RECORD = 'periodDailyRecord',
}